NP-14998: Restrict app admin
App admin only full access on creation of one employment
