+++ cristin-person/src/main/java/no/unit/nva/cristin/person/employment/create/CreatePersonEmploymentHandler.java
@@ -1,5 +1,6 @@
 package no.unit.nva.cristin.person.employment.create;
 
+import static no.unit.nva.cristin.common.Utils.CAN_UPDATE_ANY_INSTITUTION;
 import static no.unit.nva.cristin.common.Utils.extractCristinInstitutionIdentifier;
 import static no.unit.nva.cristin.common.Utils.getValidPersonId;
 import static no.unit.nva.cristin.model.Constants.DEFAULT_RESPONSE_MEDIA_TYPES;
@@ -13,6 +14,7 @@ import no.unit.nva.utils.AccessUtils;
 import nva.commons.apigateway.ApiGatewayHandler;
 import nva.commons.apigateway.RequestInfo;
 import nva.commons.apigateway.exceptions.ApiGatewayException;
+import nva.commons.apigateway.exceptions.BadRequestException;
 import nva.commons.apigateway.exceptions.ForbiddenException;
 import nva.commons.core.Environment;
 import nva.commons.core.JacocoGenerated;
@@ -40,9 +42,10 @@ public class CreatePersonEmploymentHandler extends ApiGatewayHandler<Employment,
         validateHasAccessRights(requestInfo);
 
         CreatePersonEmploymentValidator.validate(input);
-        String identifier = getValidPersonId(requestInfo);
+        var identifier = getValidPersonId(requestInfo);
+        var institutionIdentifier = fullAccessValueOrInstitutionNumber(requestInfo);
 
-        return apiClient.createEmploymentInCristin(identifier, input, extractCristinInstitutionIdentifier(requestInfo));
+        return apiClient.createEmploymentInCristin(identifier, input, institutionIdentifier);
     }
 
     @Override
@@ -56,8 +59,15 @@ public class CreatePersonEmploymentHandler extends ApiGatewayHandler<Employment,
     }
 
     private void validateHasAccessRights(RequestInfo requestInfo) throws ForbiddenException {
-        if (!AccessUtils.requesterIsUserAdministrator(requestInfo)) {
+        if (!AccessUtils.requesterIsUserAdministrator(requestInfo) && !requestInfo.userIsApplicationAdmin()) {
             throw new ForbiddenException();
         }
+    }
+
+    private String fullAccessValueOrInstitutionNumber(RequestInfo requestInfo)
+        throws BadRequestException, ForbiddenException {
+
+        return requestInfo.userIsApplicationAdmin()
+                   ? CAN_UPDATE_ANY_INSTITUTION : extractCristinInstitutionIdentifier(requestInfo);
     }
 }
